requirements.txt
alembic==1.7.7
atomicwrites==1.4.1
attrs==22.2.0
click==8.0.4
colorama==0.4.5
dataclasses==0.8
Flask==2.0.3
Flask-Migrate==4.0.1
Flask-SQLAlchemy==2.5.1
greenlet==2.0.2
importlib-metadata==4.8.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.1.4
pytest==7.0.1
python-dotenv==0.20.0
SQLAlchemy==1.4.54
tomli==1.2.3
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
psycopg2-binary
